+++ node/config.js
@@ -40,7 +40,8 @@ exports.getConfig = async (configFile) => {
   if (!config.envs) config.envs = []
 
   // TODO: Migrate this to its own yaml file
-  config.base.stateFiles.push('.orderFormDebug.json')
+  config.base.stateFiles.push('_orderFormDebug.json')
+  config.base.keepStateFiles = true
 
   // Run in different mode if on GitHub (CI)
   if (system.isCI()) {